+++ src/index.js
@@ -26,6 +26,8 @@ import sitemap from './sitemap/handler.js';
 import backlinks from './backlinks/handler.js';
 import experimentation from './experimentation/handler.js';
 import conversion from './conversion/handler.js';
+import essExperimentationMonthly from './experimentation-ess/monthly.js';
+import essExperimentationDaily from './experimentation-ess/daily.js';
 
 const HANDLERS = {
   apex,
@@ -37,6 +39,8 @@ const HANDLERS = {
   'broken-backlinks': backlinks,
   experimentation,
   conversion,
+  'experimentation-ess-monthly': essExperimentationMonthly,
+  'experimentation-ess-daily': essExperimentationDaily,
 };
 
 /**